requirements.txt
# Streamlit app + ML runtime (loose pins to avoid Python-version conflicts)
streamlit>=1.25.0

# CLIP + Torch (loose, prefer >=2.9)
open-clip-torch>=2.0.0
torch>=2.9.0

# YOLO (ultralytics)
ultralytics>=8.0.0

# Common libs
pillow>=9.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.2
requests>=2.28.0
